+++ src/context/param_init.c
@@ -6,7 +6,7 @@
 /*   By: rorollin <rorollin@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/10/01 14:58:20 by rorollin          #+#    #+#             */
-/*   Updated: 2025/10/01 16:01:50 by rorollin         ###   ########.fr       */
+/*   Updated: 2025/10/03 07:17:47 by rorollin         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,6 +18,6 @@ t_param	param_init(t_input input)
 
 	param.time_const = time_init(input);
 	param.philo_max = input.philo_max;
-	param.max_meal = input.max_meal;
+	param.max_meal = (int) input.max_meal;
 	return (param);
 }
